Update upgrade procedure to remove old language files - refs #4467

+++ main/install/update-files-1.9.0-1.10.0.inc.php
@@ -22,15 +22,88 @@ if (defined('SYSTEM_INSTALLATION')) {
 
     // Delete the "chat" file in all language directories, as variables have been moved to the trad4all file
     $langPath = api_get_path(SYS_CODE_PATH).'lang/';
+    // Only erase files from Chamilo languages (not sublanguages defined by the users)
+    $officialLanguages = array(
+        'arabic',
+        'asturian',
+        'basque',
+        'bengali',
+        'bosnian',
+        'brazilian',
+        'bulgarian',
+        'catalan',
+        'croatian',
+        'czech',
+        'danish',
+        'dari',
+        'dutch',
+        'english',
+        'esperanto',
+        'faroese',
+        'finnish',
+        'french',
+        'friulian',
+        'galician',
+        'georgian',
+        'german',
+        'greek',
+        'hebrew',
+        'hindi',
+        'hungarian',
+        'indonesian',
+        'italian',
+        'japanese',
+        'korean',
+        'latvian',
+        'lithuanian',
+        'macedonian',
+        'malay',
+        'norwegian',
+        'occitan',
+        'pashto',
+        'persian',
+        'polish',
+        'portuguese',
+        'quechua_cusco',
+        'romanian',
+        'russian',
+        'serbian',
+        'simpl_chinese',
+        'slovak',
+        'slovenian',
+        'somali',
+        'spanish',
+        'spanish_latin',
+        'swahili',
+        'swedish',
+        'tagalog',
+        'thai',
+        'tibetan',
+        'trad_chinese',
+        'turkish',
+        'ukrainian',
+        'vietnamese',
+        'xhosa',
+        'yoruba',
+    );
     $list = scandir($langPath);
     foreach ($list as $entry) {
-        if (is_dir($langPath.$entry)) {
+        if (is_dir($langPath.$entry) && in_array($entry, $officialLanguages)) {
             unlink($langPath.$entry.'/accessibility.inc.php');
             unlink($langPath.$entry.'/chat.inc.php');
             unlink($langPath.$entry.'/course_description.inc.php');
+            unlink($langPath.$entry.'/course_home.inc.php');
             unlink($langPath.$entry.'/external_module.inc.php');
+            unlink($langPath.$entry.'/glossary.inc.php');
+            unlink($langPath.$entry.'/import.inc.php');
+            unlink($langPath.$entry.'/md_mix.inc.php');
+            unlink($langPath.$entry.'/messages.inc.php');
             unlink($langPath.$entry.'/myagenda.inc.php');
+            unlink($langPath.$entry.'/notebook.inc.php');
+            unlink($langPath.$entry.'/pedaSuggest.inc.php');
+            unlink($langPath.$entry.'/resourcelinker.inc.php');
             unlink($langPath.$entry.'/scormbuilder.inc.php');
+            unlink($langPath.$entry.'/scormdocument.inc.php');
             unlink($langPath.$entry.'/slideshow.inc.php');
         }
     }
